Add files via upload
# Project Overview

Understanding If, Else Statements: These are crucial in programming, as they enable the programs you write to make decisions on their own. 
By using these statements, your program can decide to execute certain portions of code based on specific conditions.

Elif Statement: Elif (else if) comes in when you need to check multiple conditions and execute different code blocks for each condition.
With the while loop, you'll learn how to create a loop that repeats as long as a certain condition is true. 
This is perfect for scenarios where you need to iterate until a specific condition is met.

The for loop, on the other hand, is ideal for situations where you already know or can programmatically determine how many times you want to repeat an action. 
Whether iterating over a sequence of elements or performing a set number of iterations, the for loop has got you covered.

---

### Requirements and Expectations

Create a Simulation

---

## Guidelines

Instructions:

Our task is to create a simple simulation program with the theme of our choice. 
This will be a text-based simulation program where the user is presented with a situation and a set of options. 
Depending on what they choose, they'll be led down different paths. 

The game should be at least three stages deep, i.e., there should be three levels of decisions to be made.

At the end, ask the user if they would like to run the sim again and replay the sim for them based on their choice.

You can choose to create an adventure in a haunted house, jungle expedition, alien planet, or any other fun scenario you can think of. 
Just be sure to use at least 3 sets of if, elif, and else statements, along with at least 2 loops!

Loops will save you from rewriting many lines of code, allowing us to automate repetitive tasks and make our programs more efficient. 
Imagine being able to execute a block of code multiple times without having to write it over and over again. That's precisely what loops enable us to do!

Here's a small example of an output of a [Moonbase Sim](https://synchronic.uat.edu/courses/6012/files/918740/preview)